pyproject.toml
[project]
name = "resoul"
version = "0.1.0"
description = "cli application for doing tts using openvoice and or xttsv2"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # additionally added for trying new preprocessing
    'spacy',
    # naively left as is from OpenVoice
    # 'librosa==0.9.1',
    'librosa',
    'faster-whisper==0.9.0',
    'pydub==0.25.1',
    'wavmark==0.0.3',
    'numpy==1.22.0',
    # 'numpy==1.26.4',
    'numpy',
    'eng_to_ipa==0.0.2',
    # 'inflect==7.0.0',
    'inflect',
    # 'unidecode==1.3.8',
    'unidecode',
    'whisper-timestamped==1.14.2',
    'pypinyin==0.50.0',
    'cn2an==0.5.22',
    'jieba==0.42.1',
    'gradio==3.48.0',
    'langid==1.1.6',
    'TTS',
    'python-dotenv'
]
